make nicer parking space page loaders
