+++ packages/frontend/src/pages/ParkingSpace/components/index.ts
@@ -1,2 +1,5 @@
-export * from './ParkingSpaceLoading'
 export * from './RemoveApplicantFromListing'
+export * from './Applicants'
+export * from './ParkingSpaceInfo'
+export * from './ApplicantsLoading'
+export * from './ParkingSpaceInfoLoading'
